Organize tests into two types
There are two types of test cases here: tricky URLs and web platform
tests. The tricky URLs are just meant to be loaded in a web browser to
see how the main URL display handles them. The web platform tests
exercise some feature of the web platform (like a permission request)
to see how that feature handles the tricky URL. This PR separates the
tests and adds some documentation around them to make it more clear
how they're meant to be used.